pyproject.toml
[project]
name = "dbt-tutorial"
version = "0.1.0"
description = "Tutorial project to show how to get started with dbt"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dbt>=1.0.0.38.12",
    "dbt-duckdb>=1.8.3",
]
